EVG-15281 Move BB validation to modify projet route
