+++ plugin/build_baron_test.go
@@ -11,103 +11,6 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestValidateBbProject(t *testing.T) {
-	assert := assert.New(t)
-	assert.NoError(db.ClearCollections(model.ProjectRefCollection))
-	p := model.ProjectRef{
-		Identifier: "proj1",
-	}
-	assert.NoError(p.Insert())
-	assert.True(validateBbProject("proj1", evergreen.BuildBaronSettings{
-		TicketCreateProject:  "BFG",
-		TicketSearchProjects: []string{"BF", "BFG"},
-	}))
-
-	assert.True(validateBbProject("proj1", evergreen.BuildBaronSettings{
-		TicketCreateProject:  "BFG",
-		TicketSearchProjects: []string{"BF", "BFG"},
-	}))
-
-	assert.True(validateBbProject("proj1", evergreen.BuildBaronSettings{
-		TicketCreateProject:  "BFG",
-		TicketSearchProjects: []string{"BF", "BFG"},
-	}))
-
-	assert.False(validateBbProject("proj1", evergreen.BuildBaronSettings{
-		TicketCreateProject: "BFG",
-	}))
-
-	assert.False(validateBbProject("proj1", evergreen.BuildBaronSettings{
-		TicketSearchProjects: []string{"BF", "BFG"},
-	}))
-}
-
-func TestBuildBaronPluginConfigureBFSuggestion(t *testing.T) {
-	assert := assert.New(t)
-	assert.NoError(db.ClearCollections(model.ProjectRefCollection))
-	p := model.ProjectRef{
-		Identifier: "proj1",
-	}
-	assert.NoError(p.Insert())
-	assert.True(validateBbProject("proj1", evergreen.BuildBaronSettings{
-		TicketCreateProject:     "BFG",
-		TicketSearchProjects:    []string{"BF", "BFG"},
-		BFSuggestionServer:      "https://evergreen.mongodb.com",
-		BFSuggestionUsername:    "user",
-		BFSuggestionPassword:    "pass",
-		BFSuggestionTimeoutSecs: 10,
-	}))
-
-	assert.True(validateBbProject("proj1", evergreen.BuildBaronSettings{
-		TicketCreateProject:     "BFG",
-		TicketSearchProjects:    []string{"BF", "BFG"},
-		BFSuggestionServer:      "https://evergreen.mongodb.com",
-		BFSuggestionTimeoutSecs: 10,
-	}))
-
-	assert.False(validateBbProject("proj1", evergreen.BuildBaronSettings{
-		TicketCreateProject:  "BFG",
-		TicketSearchProjects: []string{"BF", "BFG"},
-		BFSuggestionUsername: "user",
-		BFSuggestionPassword: "pass",
-	}))
-
-	assert.False(validateBbProject("proj1", evergreen.BuildBaronSettings{
-		TicketCreateProject:     "BFG",
-		TicketSearchProjects:    []string{"BF", "BFG"},
-		BFSuggestionTimeoutSecs: 10,
-	}))
-
-	assert.False(validateBbProject("proj1", evergreen.BuildBaronSettings{
-		TicketCreateProject:     "BFG",
-		TicketSearchProjects:    []string{"BF", "BFG"},
-		BFSuggestionServer:      "://evergreen.mongodb.com",
-		BFSuggestionTimeoutSecs: 10,
-	}))
-
-	assert.False(validateBbProject("proj1", evergreen.BuildBaronSettings{
-		TicketCreateProject:     "BFG",
-		TicketSearchProjects:    []string{"BF", "BFG"},
-		BFSuggestionServer:      "https://evergreen.mongodb.com",
-		BFSuggestionPassword:    "pass",
-		BFSuggestionTimeoutSecs: 10,
-	}))
-
-	assert.False(validateBbProject("proj1", evergreen.BuildBaronSettings{
-		TicketCreateProject:     "BFG",
-		TicketSearchProjects:    []string{"BF", "BFG"},
-		BFSuggestionServer:      "https://evergreen.mongodb.com",
-		BFSuggestionTimeoutSecs: 0,
-	}))
-
-	assert.False(validateBbProject("proj1", evergreen.BuildBaronSettings{
-		TicketCreateProject:     "BFG",
-		TicketSearchProjects:    []string{"BF", "BFG"},
-		BFSuggestionServer:      "https://evergreen.mongodb.com",
-		BFSuggestionTimeoutSecs: -1,
-	}))
-}
-
 func TestBbGetProject(t *testing.T) {
 	require.NoError(t, db.ClearCollections(task.Collection, model.ProjectRefCollection, model.ParserProjectCollection),
 		"Error clearing task collections")
