+++ service/project_test.go
@@ -37,3 +37,100 @@ func TestVerifyAliasExists(t *testing.T) {
 	assert.NoError(t, err)
 	assert.False(t, exists)
 }
+
+func TestValidateBbProject(t *testing.T) {
+	assert := assert.New(t)
+	assert.NoError(db.ClearCollections(model.ProjectRefCollection))
+	p := model.ProjectRef{
+		Identifier: "proj1",
+	}
+	assert.NoError(p.Insert())
+	assert.Nil(bbProjectIsValid("proj1", evergreen.BuildBaronSettings{
+		TicketCreateProject:  "BFG",
+		TicketSearchProjects: []string{"BF", "BFG"},
+	}))
+
+	assert.Nil(bbProjectIsValid("proj1", evergreen.BuildBaronSettings{
+		TicketCreateProject:  "BFG",
+		TicketSearchProjects: []string{"BF", "BFG"},
+	}))
+
+	assert.Nil(bbProjectIsValid("proj1", evergreen.BuildBaronSettings{
+		TicketCreateProject:  "BFG",
+		TicketSearchProjects: []string{"BF", "BFG"},
+	}))
+
+	assert.NotNil(bbProjectIsValid("proj1", evergreen.BuildBaronSettings{
+		TicketCreateProject: "BFG",
+	}))
+
+	assert.NotNil(bbProjectIsValid("proj1", evergreen.BuildBaronSettings{
+		TicketSearchProjects: []string{"BF", "BFG"},
+	}))
+}
+
+func TestBuildBaronPluginConfigureBFSuggestion(t *testing.T) {
+	assert := assert.New(t)
+	assert.NoError(db.ClearCollections(model.ProjectRefCollection))
+	p := model.ProjectRef{
+		Identifier: "proj1",
+	}
+	assert.NoError(p.Insert())
+	assert.Nil(bbProjectIsValid("proj1", evergreen.BuildBaronSettings{
+		TicketCreateProject:     "BFG",
+		TicketSearchProjects:    []string{"BF", "BFG"},
+		BFSuggestionServer:      "https://evergreen.mongodb.com",
+		BFSuggestionUsername:    "user",
+		BFSuggestionPassword:    "pass",
+		BFSuggestionTimeoutSecs: 10,
+	}))
+
+	assert.Nil(bbProjectIsValid("proj1", evergreen.BuildBaronSettings{
+		TicketCreateProject:     "BFG",
+		TicketSearchProjects:    []string{"BF", "BFG"},
+		BFSuggestionServer:      "https://evergreen.mongodb.com",
+		BFSuggestionTimeoutSecs: 10,
+	}))
+
+	assert.NotNil(bbProjectIsValid("proj1", evergreen.BuildBaronSettings{
+		TicketCreateProject:  "BFG",
+		TicketSearchProjects: []string{"BF", "BFG"},
+		BFSuggestionUsername: "user",
+		BFSuggestionPassword: "pass",
+	}))
+
+	assert.NotNil(bbProjectIsValid("proj1", evergreen.BuildBaronSettings{
+		TicketCreateProject:     "BFG",
+		TicketSearchProjects:    []string{"BF", "BFG"},
+		BFSuggestionTimeoutSecs: 10,
+	}))
+
+	assert.NotNil(bbProjectIsValid("proj1", evergreen.BuildBaronSettings{
+		TicketCreateProject:     "BFG",
+		TicketSearchProjects:    []string{"BF", "BFG"},
+		BFSuggestionServer:      "://evergreen.mongodb.com",
+		BFSuggestionTimeoutSecs: 10,
+	}))
+
+	assert.NotNil(bbProjectIsValid("proj1", evergreen.BuildBaronSettings{
+		TicketCreateProject:     "BFG",
+		TicketSearchProjects:    []string{"BF", "BFG"},
+		BFSuggestionServer:      "https://evergreen.mongodb.com",
+		BFSuggestionPassword:    "pass",
+		BFSuggestionTimeoutSecs: 10,
+	}))
+
+	assert.NotNil(bbProjectIsValid("proj1", evergreen.BuildBaronSettings{
+		TicketCreateProject:     "BFG",
+		TicketSearchProjects:    []string{"BF", "BFG"},
+		BFSuggestionServer:      "https://evergreen.mongodb.com",
+		BFSuggestionTimeoutSecs: 0,
+	}))
+
+	assert.NotNil(bbProjectIsValid("proj1", evergreen.BuildBaronSettings{
+		TicketCreateProject:     "BFG",
+		TicketSearchProjects:    []string{"BF", "BFG"},
+		BFSuggestionServer:      "https://evergreen.mongodb.com",
+		BFSuggestionTimeoutSecs: -1,
+	}))
+}
